requirements/dev.txt
-r base.txt

django-debug-toolbar==4.4.2